Renamed file structure to match example better
