+++ client/src/main/java/org/sepia/skelly/client/SkellyClientModule.java
@@ -1,7 +1,7 @@
 package org.sepia.skelly.client;
 
-import org.sepia.skelly.common.AbstractSkellyModule;
-import org.sepia.skelly.common.SkellyMessage;
+import org.sepia.skelly.scripting.AbstractSkellyModule;
+import org.sepia.skelly.scripting.SkellyMessage;
 
 import com.inductiveautomation.ignition.client.gateway_interface.ModuleRPCFactory;
