+++ common/src/main/resources/org/sepia/skelly/scripting/AbstractSkellyModule.properties
@@ -1,0 +1,2 @@
+HelloWorld.desc=A nice greeting
+HelloWorld.returns=Returns a message
