+++ designer/build.gradle
@@ -1,24 +1,14 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This project uses @Incubating APIs which are subject to change.
- */
-
 plugins {
-   id "java-library"
+    id "java-library"
 }
 
 sourceCompatibility = JavaVersion.VERSION_11
 targetCompatibility = JavaVersion.VERSION_11
 
 dependencies {
-    compile(project(':client'))
-    compile(project(':common'))
+    compile(project(":common"))
 
-    compileOnly 'com.inductiveautomation.ignitionsdk:ignition-common:8.1.0-SNAPSHOT'
-    compileOnly 'com.inductiveautomation.ignitionsdk:client-api:8.1.0-SNAPSHOT'
-    compileOnly 'com.inductiveautomation.ignitionsdk:designer-api:8.1.0-SNAPSHOT'
-    compileOnly 'com.inductiveautomation.ignitionsdk:vision-designer-api:8.1.0-SNAPSHOT'
+    compileOnly("com.inductiveautomation.ignitionsdk:ignition-common:${sdk_version}")
+    compileOnly("com.inductiveautomation.ignitionsdk:client-api:${sdk_version}")
+    compileOnly("com.inductiveautomation.ignitionsdk:designer-api:${sdk_version}")    
 }
-
-description = 'designer'
