+++ designer/src/main/java/org/sepia/skelly/designer/SkellyDesignerHook.java
@@ -1,6 +1,7 @@
-package org.sepia.skelly;
+package org.sepia.skelly.designer;
 
 import org.sepia.skelly.client.SkellyClientModule;
+
 import com.inductiveautomation.ignition.common.script.ScriptManager;
 import com.inductiveautomation.ignition.common.script.hints.PropertiesFileDocProvider;
 import com.inductiveautomation.ignition.designer.model.AbstractDesignerModuleHook;
